chore(SearchModule): reorganize file structure
