+++ src/app/search/components/search-bar.scss
