+++ src/app/search/components/search-bar.spec.ts
