+++ src/app/search/components/search-bar.ts
