+++ src/app/search/index.ts
@@ -1,0 +1,4 @@
+export { searchReducer, SearchState } from './state';
+export { SearchActions } from './search-actions';
+export { SearchModule } from './search-module';
+export { SearchService } from './search-service';
