+++ src/app/search/state/index.ts
@@ -1,0 +1,2 @@
+export { searchReducer } from './search-reducer';
+export { SearchState } from './search-state';
