+++ src/app/search/state/search-reducer.spec.ts
