+++ src/app/search/state/search-reducer.ts
