+++ src/tracklists/reducers/tracklist-reducer.spec.ts
@@ -1,6 +1,6 @@
 import { is, List } from 'immutable';
 import { TRACKS_PER_PAGE } from 'src/constants';
-import { SearchActions } from 'src/search/search-actions';
+import { SearchActions } from 'src/app/search/search-actions';
 import { testUtils } from 'src/utils/test';
 import { Tracklist, TracklistRecord } from '../models/tracklist';
 import { TracklistActions } from '../tracklist-actions';
