+++ src/tracklists/reducers/tracklist-reducer.ts
@@ -1,7 +1,7 @@
 import { Action, ActionReducer } from '@ngrx/store';
 import { List } from 'immutable';
 import { TRACKS_PER_PAGE } from 'src/constants';
-import { SearchActions } from 'src/search/search-actions';
+import { SearchActions } from 'src/app/search/search-actions';
 import { UserActions } from 'src/users/user-actions';
 import { TrackData } from '../models/track';
 import { Tracklist, TracklistRecord } from '../models/tracklist';
