+++ src/tracklists/reducers/tracklists-reducer.ts
@@ -1,6 +1,6 @@
 import { Action, ActionReducer } from '@ngrx/store';
 import { Map } from 'immutable';
-import { SearchActions } from 'src/search/search-actions';
+import { SearchActions } from 'src/app/search/search-actions';
 import { UserActions } from 'src/users/user-actions';
 import { TracklistActions } from '../tracklist-actions';
 import { tracklistReducer } from './tracklist-reducer';
